requirements.txt
streamlit==1.32.0
pandas==2.0.3
nltk==3.8.1
scikit-learn==1.4.0
PyPDF2==3.0.1
pdfminer.six==20250506
python-docx==1.1.0 